requirements.txt
python-docx
openpyxl
pytest
pydantic       
sqlalchemy  
